moving SSLContext and SSLSocket to __init__ to avoid ciclic includes
